Move Threshold and coexisting implementations into packages/utils
